Moved security classes in podium-common to common namespace.
