+++ podium-common/src/main/java/org/bbmri/podium/common/security/annotations/AnyAuthorisedUser.java
@@ -5,22 +5,15 @@
  * See the file LICENSE in the root of this repository.
  */
 
-package org.bbmri.podium.security.annotations;
+package org.bbmri.podium.common.security.annotations;
 
 import java.lang.annotation.*;
 
 /**
- * Mark a controller method as accessible for any authorised user with
- * any of the authorities in {@link #value()}.
+ * Mark a controller method as accessible for any authorised user,
+ * i.e., any user that is logged in or has a provided a valid authorisation token.
  */
 @Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface SecuredByAuthority {
-
-    /**
-     * Authority names. See {@link org.bbmri.podium.domain.Authority}.
-     */
-    String[] value() default {};
-
-}
+public @interface AnyAuthorisedUser {}
