+++ podium-gateway/src/main/java/org/bbmri/podium/config/MicroserviceSecurityConfiguration.java
@@ -10,7 +10,7 @@
 
 package org.bbmri.podium.config;
 
-import org.bbmri.podium.security.AuthorityConstants;
+import org.bbmri.podium.common.security.AuthorityConstants;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.cloud.client.loadbalancer.RestTemplateCustomizer;
