--- podium-gateway/src/main/java/org/bbmri/podium/security/AuthorityConstants.java
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2017. The Hyve and respective contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- *
- * See the file LICENSE in the root of this repository.
- *
- */
-
-package org.bbmri.podium.security;
-
-public final class AuthorityConstants {
-
-    public static final String ANONYMOUS = "ROLE_ANONYMOUS";
-
-    public static final String PODIUM_ADMIN                 = "ROLE_PODIUM_ADMIN";
-    public static final String BBMRI_ADMIN                  = "ROLE_BBMRI_ADMIN";
-    public static final String ORGANISATION_ADMIN           = "ROLE_ORGANISATION_ADMIN";
-    public static final String ORGANISATION_COORDINATOR     = "ROLE_ORGANISATION_COORDINATOR";
-    public static final String REVIEWER                     = "ROLE_REVIEWER";
-    public static final String RESEARCHER                   = "ROLE_RESEARCHER";
-
-}
