+++ podium-gateway/src/main/java/org/bbmri/podium/security/SecurityUtils.java
@@ -10,6 +10,7 @@
 
 package org.bbmri.podium.security;
 
+import org.bbmri.podium.common.security.AuthorityConstants;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
