+++ podium-uaa/src/main/java/org/bbmri/podium/domain/Authority.java
@@ -10,7 +10,7 @@
 
 package org.bbmri.podium.domain;
 
-import org.bbmri.podium.security.AuthorityConstants;
+import org.bbmri.podium.common.security.AuthorityConstants;
 import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
 import javax.persistence.Entity;
@@ -20,8 +20,6 @@ import javax.persistence.Column;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.io.Serializable;
-import java.util.HashSet;
-import java.util.Set;
 
 /**
  * An authority (a security role) used by Spring Security.
