+++ podium-uaa/src/main/java/org/bbmri/podium/security/SecurityService.java
@@ -10,7 +10,8 @@
 
 package org.bbmri.podium.security;
 
-import org.bbmri.podium.domain.Authority;
+import org.bbmri.podium.common.security.AuthorityConstants;
+import org.bbmri.podium.common.security.UserAuthenticationToken;
 import org.bbmri.podium.domain.User;
 import org.bbmri.podium.service.UserService;
 import org.slf4j.Logger;
