+++ podium-uaa/src/main/java/org/bbmri/podium/web/rest/RoleResource.java
@@ -12,12 +12,11 @@ package org.bbmri.podium.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
 import com.google.common.collect.Sets;
-import org.bbmri.podium.security.AuthorityConstants;
-import org.bbmri.podium.security.annotations.OrganisationParameter;
-import org.bbmri.podium.security.annotations.OrganisationUuidParameter;
-import org.bbmri.podium.security.annotations.SecuredByOrganisation;
-import org.bbmri.podium.security.annotations.SecuredByAuthority;
-import org.bbmri.podium.domain.Authority;
+import org.bbmri.podium.common.security.AuthorityConstants;
+import org.bbmri.podium.common.security.annotations.OrganisationParameter;
+import org.bbmri.podium.common.security.annotations.OrganisationUuidParameter;
+import org.bbmri.podium.common.security.annotations.SecuredByOrganisation;
+import org.bbmri.podium.common.security.annotations.SecuredByAuthority;
 import org.bbmri.podium.domain.Organisation;
 import org.bbmri.podium.domain.Role;
 import org.bbmri.podium.domain.User;
