+++ podium-uaa/src/test/java/org/bbmri/podium/security/OAuth2TokenMockUtil.java
@@ -10,6 +10,7 @@
 
 package org.bbmri.podium.security;
 
+import org.bbmri.podium.common.security.UserAuthenticationToken;
 import org.bbmri.podium.domain.User;
 import org.mockito.internal.util.collections.Sets;
 import org.springframework.boot.test.mock.mockito.MockBean;
