+++ podium-uaa/src/test/java/org/bbmri/podium/web/rest/AuthenticationIntTest.java
@@ -8,9 +8,8 @@
 package org.bbmri.podium.web.rest;
 
 import org.bbmri.podium.PodiumUaaApp;
-import org.bbmri.podium.domain.Authority;
 import org.bbmri.podium.domain.User;
-import org.bbmri.podium.security.AuthorityConstants;
+import org.bbmri.podium.common.security.AuthorityConstants;
 import org.bbmri.podium.security.OAuth2TokenMockUtil;
 import org.bbmri.podium.service.UserService;
 import org.bbmri.podium.web.rest.vm.ManagedUserVM;
